requirements.txt
requests
reportlab
python-dotenv
openai
aiohttp
